Add a zero transformation which maps input gradients to zero.
As mentioned in issue #175, which proposed this feature, a zero transformation
is useful for example in combination with multi_transform as it allows keeping
certain parameters fixed when the updates are applied.

This PR also adds a note to the docstring of the identity transform to avoid
confusion with the zero_transform. The former leaves gradients unchanged, the
latter maps gradients to zero so that the parameters are unchanged when the
gradient-based updates are applied.

PiperOrigin-RevId: 405162051